Refactor: 반복되는 조회 함수 분리 및 여러개의 인자를 하나의 객체로 수정
